app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic==2.7.1
pydantic-settings==2.2.1
starlette==0.36.3
typing-extensions>=4.10.0